Fix watched tubes not restored after reconnect.
Due to the way Beaneater::Connection#transmit handles it parameters,
`_initialize_tubes` was never called.

This causes beaneater to watch the `default` tube whenever it
silently reconnects to server.

For example, a client watches a tube named `jobs`, while it waits for
a job using `reserve`, a connection failure caused beaneater to reconnect.

Since reconnect is transparent to the client,
the client's reserve command will block indefinitely, as the watched
tube has been resetted to `default` after reconnect.

This commit also fixes a bug in `_initialize_tubes`,
where `default` tube wasn't ignored when it's not in the watch list.